feat: multi-user task assignments, meetings participants, leave/security fixes
## Core architecture
- Replace tasks.assigned_to (single-user) with task_assignments table
  - One task → N assignments, each with independent status lifecycle
  - assigned_to column dropped from DB (migration scripts included)
  - _sync_task_aggregate_status() keeps task.status consistent
  - joinedload replaced with explicit JOIN queries (no tasks_1 alias)

## Task system
- Task creation: single path for 1 user, bulk path for N users
- Employee view: assignments-based fetch (list_tasks_for_employee)
- Manager view: one row per assignee with per-user approve/reject
- Status lifecycle: assigned → in_progress → pending_approval → completed
- Filters: title, employee name, date range (server-side)
- Progress panel: grouped by status with scrollable chip list
- Mobile responsive: card layout, stacked assignees

## Meetings
- New MeetingParticipant model (meeting_participants table)
- Multi-select UI with filter/select-all/clear helpers
- Batch participant load (no N+1); legacy participant_id preserved
- Notifications sent to all participants on create

## Leave management
- Block self-approval: backend 403 guard + all 4 template conditions

## Notification system
- Per-module unread counts (sidebar badges + bell)
- WebSocket route for real-time updates
- N+1 fix: batch-load user emails before notification loop

## Security & validation
- CSV/JSON export: guard nullable .value on status/priority
- Meetings form: Form(default=[]) → 400 instead of 422 on empty
- Hierarchy checks applied consistently across task/leave/meeting scopes

## DB migrations (scripts/)
- migrate_assigned_to.py: make assigned_to nullable + backfill assignments
- drop_assigned_to.py: DROP COLUMN after backfill verified
- migrate_batch_id.py: batch_id column for legacy grouping

+++ app/models/expense.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, UniqueConstraint
+from sqlalchemy import Column, Integer, String, Numeric, DateTime, ForeignKey, UniqueConstraint
 
 from app.core.database import Base
 
@@ -13,7 +13,7 @@ class ExpenseGroup(Base):
     id           = Column(Integer, primary_key=True, index=True)
     title        = Column(String(200), nullable=False)
     created_by   = Column(Integer, ForeignKey("users.id"), nullable=False)
-    total_amount = Column(Float, nullable=False)
+    total_amount = Column(Numeric(10, 2), nullable=False)  # max 99,999,999.99
     created_at   = Column(DateTime, default=datetime.utcnow, nullable=False)
 
 
@@ -26,7 +26,7 @@ class ExpenseMember(Base):
     group_id      = Column(Integer, ForeignKey("expense_groups.id", ondelete="CASCADE"),
                            nullable=False, index=True)
     user_id       = Column(Integer, ForeignKey("users.id"), nullable=False)
-    amount_share  = Column(Float, nullable=False)
+    amount_share  = Column(Numeric(10, 2), nullable=False)  # max 99,999,999.99
     status        = Column(String(20), default="pending", nullable=False)  # pending | paid
 
     __table_args__ = (
